requirements.txt
feedparser
requests
beautifulsoup4
anthropic
python-telegram-bot==13.15
apscheduler
